requirements.txt
anyjson==0.3.3
backports.csv==1.0.7
diff-match-patch==20181111
dj-database-url==0.4.0
Django==1.11.17
django-admin-tools==0.7.2
django-braces==1.8.1
django-celery==3.1.17
django-import-export==1.2.0
et-xmlfile==1.0.1
gunicorn==19.4.5
jdcal==1.4.1
jsonfield==1.0.3
MarkupPy==1.14
nsetools==1.0.1
odfpy==1.4.1
openpyxl==2.6.4
pytz==2019.3
PyYAML==5.3.1
redis==3.2.0
six==1.14.0
tablib==0.14.0
whitenoise==2.0.6
xlrd==1.2.0
xlwt==1.3.0
ystockquote==0.2.4
mysqlclient==1.3.12
django-mysql==2.2.0
